More UI conveniences.
Close tabs with middle mouse button.

Image options show a thumbnail of the selected image.

Image picker shows name of image and actually wraps when there are too
many images, instead of horizontally scrolling. A double click anywhere
inside the margin of an image will select it, instead of only directly
on the image element. The currently active image in the main window is
selected by default.